We now have custom modal submit buttons.
They can be accessed with a widget call to ModalSubmitButton.
These buttons should be placed in forms.

The buttons take two parameters in an array format similar to other widgets.

label is the label you wish to put on the button.

modelName is the name of the form you are submitting.