Abstracted the details of the modal open button.
This button can be called with ModalOpenButton. It opens up a modal window
for the given url.

This button takes up to four parameters in the standard widget format.

label is a label for the button.

url is the url of page you wish the modal to open. You must ensure that
renderModal is in the controller for that action.

button_id is the unique id for the button itself.

type is optional, and it takes one of the types listed in TbButton.