Extra modal submits fixed for ModalSubmitButton.
Modals the submit using the ModalSubmitButton no longer submit multiple
times after failing once. Each modal submit button now has a unique id on
top of a id.

It is also no longer required to have a model name for the submit button.
ModelName is now deprecated for the ModalSubmitButton.